+++ Assets/Scripts/Incidents/Contexts/Organizations/OrganizationPosition.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d67de383612f1394f9999e8330cfa566
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
